update licence as per https://github.com/librestack/librecast/issues/39
As the Free Software Foundation has decided to reinstate Richard Stallman, The Librecast project no longer feels that the FSF is an appropriate steward of Free Software and wish to ensure that any future versions of the GPL do not automatically apply to Librecast.

Therefore we have decided to drop the "or later" part of the GPL licence. We have also decided to dual licence Librecast and add GPLv2.

We still believe in the GPL and copyleft, however the Free Software Movement is more than The Free Software Foundation. Like any other project we have benefited from the four freedoms and we believe in them. We are also free to demonstrate our current lack of trust in the board of the FSF.